+++ notificare-geo/src/main/java/re/notifica/geo/internal/storage/LocalStorage.kt
@@ -321,4 +321,10 @@ internal class LocalStorage(context: Context) {
     }
 
     // endregion
+
+    fun clear() {
+        sharedPreferences.edit {
+            clear()
+        }
+    }
 }