+++ notificare-inbox/src/main/java/re/notifica/inbox/internal/NotificareInboxImpl.kt
@@ -86,6 +86,10 @@ internal object NotificareInboxImpl : NotificareModule(), NotificareInbox {
         reloadLiveItems()
     }
 
+    override suspend fun clearStorage() {
+        database.inbox().clear()
+    }
+
     override suspend fun launch() {
         sync()
     }