+++ notificare-push/src/main/java/re/notifica/push/NotificareFirebaseMessagingService.kt
@@ -7,12 +7,11 @@ import re.notifica.push.internal.NotificareNotificationRemoteMessage
 import re.notifica.push.internal.NotificareSystemRemoteMessage
 import re.notifica.push.internal.NotificareUnknownRemoteMessage
 import re.notifica.push.internal.logger
-import re.notifica.push.ktx.isNotificareNotification
 import re.notifica.push.ktx.push
 import re.notifica.push.ktx.pushInternal
 import re.notifica.push.models.NotificareTransport
 
-public open class NotificarePushService : FirebaseMessagingService() {
+public open class NotificareFirebaseMessagingService : FirebaseMessagingService() {
 
     override fun onNewToken(token: String) {
         Notificare.pushInternal().handleNewToken(NotificareTransport.GCM, token)
@@ -22,6 +21,17 @@ public open class NotificarePushService : FirebaseMessagingService() {
         logger.debug("Received a remote notification from FCM.")
 
         if (Notificare.push().isNotificareNotification(message)) {
+            val application = Notificare.application ?: run {
+                @Suppress("detekt:MaxLineLength")
+                NotificareLogger.warning("Notificare application unavailable. Ensure Notificare is configured during the application launch.")
+                return
+            }
+
+            if (application.id != message.data["x-application"]) {
+                NotificareLogger.warning("Incoming notification originated from another application.")
+                return
+            }
+
             val isSystemNotification = message.data["system"] == "1" ||
                 message.data["system"]?.toBoolean() ?: false