+++ notificare-push/src/main/java/re/notifica/push/NotificarePush.kt
@@ -2,8 +2,11 @@ package re.notifica.push
 
 import android.content.Intent
 import androidx.lifecycle.LiveData
+import com.google.firebase.messaging.RemoteMessage
 import re.notifica.InternalNotificareApi
 import re.notifica.NotificareCallback
+import re.notifica.push.models.NotificareNotificationActionOpenedIntentResult
+import re.notifica.push.models.NotificareNotificationOpenedIntentResult
 import re.notifica.push.models.NotificarePushSubscription
 import re.notifica.push.models.NotificareTransport
 import re.notifica.push.models.NotificareRemoteMessage
@@ -32,10 +35,13 @@ public interface NotificarePush {
 
     public fun disableRemoteNotifications(callback: NotificareCallback<Unit>)
 
-    // Augmented in the appropriate peer module.
-    // public fun isNotificareNotification(...: RemoteMessage)
+    public fun isNotificareNotification(remoteMessage: RemoteMessage): Boolean
 
     public fun handleTrampolineIntent(intent: Intent): Boolean
+
+    public fun parseNotificationOpenedIntent(intent: Intent): NotificareNotificationOpenedIntentResult?
+
+    public fun parseNotificationActionOpenedIntent(intent: Intent): NotificareNotificationActionOpenedIntentResult?
 
     public suspend fun registerLiveActivity(activityId: String, topics: List<String> = listOf())