+++ notificare/src/main/java/re/notifica/internal/storage/database/dao/NotificareEventsDao.kt
@@ -22,4 +22,7 @@ internal interface NotificareEventsDao {
 
     @Delete
     suspend fun delete(event: NotificareEventEntity)
+
+    @Query("DELETE FROM events")
+    suspend fun clear()
 }